requirements.txt
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.7.0
openpyxl>=3.1.2
pandas>=2.0.0
numpy>=1.24.0
pillow>=9.5.0
langchain>=0.0.267
python-dotenv>=1.0.0 